+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.17.0-0.7.1 (2025-05-07)
 
 * [PLT-2098] Improve kubernetes version checks during cloud-provisioner-upgrade
+* [PLT-2124] Bump cluster-autoscaler to v1.32.0 version and its chart version to 9.46.6
 
 ## 0.17.0-0.7.0 (2025-04-30)